# API/openapi.yaml
openapi: 3.0.0
info:
    title: Supplier, Warehouse, and Good Management API
    version: 1.0.0
    description: API for managing suppliers, warehouses, and goods
servers:
    - url: /api/v1 # Replace with your actual base path
components:
    schemas:
...
        error:
            type: object
            properties:
                code:
                    type: number
                title:
                    type: string
                details:
                    type: string
paths:
...
    /refreshToken:
        post:
            summary: Refresh token
            tags:
                - Auth

            responses:
                '422':
                    description: Refresh token expired or user is logged out
                    content:
                        application/json:
                            schema:
                                type: object
                                items:
                                    $ref: '#/components/schemas/error'
                            example:
                                code: 422
                                title: Unprocessable Entity
                                details: Session is expired or revoked
                '200':
                    description: Successful refresh
    /logout:
        post:
            summary: Logout
            tags:
                - Auth

            responses:
                '422':
                    description: Refresh token cookie was not found
                    content:
                        application/json:
                            schema:
                                type: object
                                items:
                                    $ref: '#/components/schemas/error'
                            example:
                                code: 422
                                title: Unprocessable Entity
                                details: Refresh token cookie was not found
                '200':
                    description: Successful logout